400 blocks of mem leak, segfault after end of exec
