+++ src/applications/wordcount.h
@@ -2,13 +2,13 @@
 
 #include "application.h"
 #include "../reader.h"
-#include "../dataframe.h"
-#include "../row.h"
+#include "../dataframe/dataframe.h"
+#include "../dataframe/row.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include "../string.h"
-#include "../key.h"
-#include "../kvstore.h"
+#include "../wrappers/string.h"
+#include "../key/key.h"
+#include "../key/kvstore.h"
 #include "../helper.h"
 #include "../array.h"
 #include "../args.h"
