+++ src/column/stringcol.h
@@ -1,37 +1,44 @@
-class StringColumn;
+/*************************************************************************
+ * StringColumn::
+ * Holds string pointers. The strings are external.  Nullptr is a valid
+ * value.
+ */
 
 class IntColumn;
+
+class BoolColumn;
 
 class FloatColumn;
 
 #pragma once
 
 #include "column.h"
-#include "intcol.h"
+#include <cstdarg>
+#include "../wrappers/string.h"
+#include "boolcol.h"
 #include "floatcol.h"
 #include "stringcol.h"
-#include "string.h"
-#include "bool.h"
+#include "intcol.h"
 #include <iostream>
 
 using namespace std;
 
 /**
- * Represent a Column of Bool
+ * Represent a Int of Float SoR Type
  */
-class BoolColumn : public Column {
+class StringColumn : public Column {
 public:
-    Bool **vals_;
+    String **vals_;
     size_t size_;
     size_t capacity_;
 
-    BoolColumn() {
+    StringColumn() {
         size_ = 0;
         capacity_ = 200 * 1000 * 1000;
-        vals_ = new Bool *[capacity_];
+        vals_ = new String *[capacity_];
     }
 
-    ~BoolColumn() {
+    ~StringColumn() {
         for (int i = 0; i < size(); i++) {
             if (vals_[i] != nullptr) {
                 delete vals_[i];
@@ -40,11 +47,19 @@ public:
         delete[] vals_;
     }
 
+    StringColumn(int n, ...) {
+        va_list args;
+        va_start(args, n);
+        for (size_t i = 0; i < n; i++) {
+            vals_[i] = new String(va_arg(args, char * ));
+        }
+    }
+
     /**
-     * Append missing bool is default 0.
-     */
+    * Append missing bool is default 0.
+    */
     void appendMissing() {
-        push_back(false);
+        this->push_back(new String(""));
     }
 
     /**
@@ -52,14 +67,14 @@ public:
      * @return
      */
     StringColumn *as_string() {
-        return nullptr;
+        return this;
     }
 
     /**
      * Returns this if it is a IntColumn
      * @return
      */
-    IntColumn *as_int() {
+    virtual IntColumn *as_int() {
         return nullptr;
     }
 
@@ -67,31 +82,28 @@ public:
      * Returns this if it is a BoolColumn
      * @return
      */
-    BoolColumn *as_bool() {
-        return this;
+    virtual BoolColumn *as_bool() {
+        return nullptr;
     }
 
     /**
      * Returns this if it is a FloatColumn
      * @return
      */
-    FloatColumn *as_float() {
+    virtual FloatColumn *as_float() {
         return nullptr;
     }
 
-    /** Returns the Bool at idx; undefined on invalid idx.*/
-    bool *get(size_t idx) {
-        if (idx >= 0 && idx <= this->size()) {
-            return &vals_[idx]->val;
-        } else {
-            exit(1);
-        }
+    /** Returns the string at idx; undefined on invalid idx.*/
+    String *get(size_t idx) {
+
+        return vals_[idx];
     }
 
     /** Out of bound idx is undefined. */
-    void set(size_t idx, bool *val) {
+    void set(size_t idx, String *val) {
         if (idx >= 0 && idx <= this->size()) {
-            vals_[idx] = new Bool(*val);
+            vals_[idx] = val;
             size_++;
         } else {
             exit(1);
@@ -99,7 +111,7 @@ public:
     }
 
     /**
-     * Returns the size of this BoolColumn
+     * Returns the size of this StringColumn
      */
     size_t size() {
         return size_;
@@ -116,8 +128,7 @@ public:
      * Adds the given bool to this if it is a BoolColumn
      */
     virtual void push_back(bool val) {
-        vals_[size_] = new Bool(val);
-        size_++;
+        exit(1);
     }
 
     /**
@@ -131,32 +142,29 @@ public:
      * Adds the given String to this if it is a StringColumn
      */
     virtual void push_back(String *val) {
-        // if passing nullptr from <MISSING> in sor then save to array as nullptr calls this method.
-        if (val == nullptr) {
-            this->vals_[size_] = nullptr;
-            size_++;
-        } else {
-            exit(1);
-        }
+        vals_[size_] = val;
+        size_++;
     }
 
     /** Return the type of this column as a char: 'S', 'B', 'I' and 'F'. */
     virtual char get_type() {
-        return 'B';
+        return 'S';
     }
 
-    /** Serializes this BoolCol **/
+    /** Returns the serialization of this StringColumn as a String */
     virtual String *serialize() {
         StrBuff *s = new StrBuff();
-        s->c("B}");
+        s->c("S}");
 
         for (int i = 0; i < this->size_; i++) {
             char str[256] = ""; /* In fact not necessary as snprintf() adds the 0-terminator. */
-            snprintf(str, sizeof str, "%d}", this->vals_[i]->val);
+            snprintf(str, sizeof str, "%s}", this->vals_[i]->c_str());
             s->c(str);
         }
 
         s->c("!");
-        return s->get();
+        String* st = s->get();
+        delete s;
+        return st;
     }
 };
