+++ src/reader.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "row.h"
+#include "dataframe/row.h"
 #include "SImap.h"
 
 class Reader {
