+++ src/rower.h
@@ -6,7 +6,7 @@
  */
 #pragma once
 
-#include "row.h"
+#include "dataframe/row.h"
 
 class Rower : public Object {
 public:
