+++ src/writer.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "row.h"
+#include "dataframe/row.h"
 #include "args.h"
 #include "SImap.h"
