+++ tests/m3/serializedf.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "../../src/network/serial.h"
-#include "../../src/dataframe.h"
+#include "../../src/dataframe/dataframe.h"
 
 int main(int argc, char* argv[]) {
